updated `Dashboard`

+++ app/src/main/java/info/fekri8614/thermocall/model/data/sensor/SensorTemperature.kt
@@ -7,7 +7,7 @@ class SensorTemperatureList : ArrayList<SensorTemperature>()
 
 data class SensorTemperature(
     @SerializedName("temperature")
-    val temperature: Int,
+    val temperature: Double,
     @SerializedName("timestamp")
     val timestamp: String
 )
